+++ camera/camera.hpp
@@ -6,13 +6,13 @@
 class Camera {
 public:
     // measured in radians
-    float yaw_angle = 0.0f;
-    float pitch_angle = 0.0f;
-    glm::vec3 look_direction = glm::vec3(0.0f, 0.0f, 1.0f);
-    glm::vec3 up_direction = glm::vec3(0.0f, 1.0f, 0.0f);
-    void update_look_direction(float change_in_yaw_angle, float change_in_pitch_angle);
+    double yaw_angle = 0.0;
+    double pitch_angle = 0.0;
+    glm::vec3 look_direction = glm::vec3(0.0, 0.0, 1.0);
+    glm::vec3 up_direction = glm::vec3(0.0, 1.0, 0.0);
+    void update_look_direction(double change_in_yaw_angle, double change_in_pitch_angle);
 private:
-    static glm::vec3 yaw_pitch_angles_to_vector(float yaw, float pitch);
+    static glm::vec3 yaw_pitch_angles_to_vector(double yaw, double pitch);
     void clamp_camera_pitch_angle();
 };
